=> had to disable the draw order since it messed up drawings a lot for transparent objects (which is worse for the playtest than the smoke-problem)

# Conflicts:
#	Project/BRS/BRS/BRS.csproj
#	Project/BRS/BRS/Content/Content.mgcb
#	Project/BRS/BRS/Content/Images/tutorial/tutorial_1.png
#	Project/BRS/BRS/Scripts/Scenes/Level1.cs

+++ Project/BRS/BRS/Engine/Particles/Settings.cs
@@ -25,7 +25,7 @@ namespace BRS.Engine.Particles {
 
 
         // Number of particles generated in each round
-        public int ParticlesPerRound = 1;
+        public float ParticlesPerRound = 1;
 
 
         // How long these particles will last.
